feat: support int and double types in buildFixedSizedBuffer

+++ types.ts
@@ -1,6 +1,6 @@
 import { EncodingOption } from "node:fs";
 
-export type DatDataTypes = "string" | "float" | "int" | "date";
+export type DatDataTypes = "string" | "double" | "int" | "date";
 
 /**
  * Represents a fixed-length entry for binary records.
